* Strict namespace setting can be a problem when upgrading existing database files, such as startup-db or persistent running-db, or any other saved XML file.   * For backward compatibility, load of startup and running set CLICON_XML_NS_STRICT to false temporarily. * Added three-valued return values for several validate functions where -1 is fatal error, 0 is validation failed and 1 is validation OK.   * This includes: `xmldb_put`, `xml_yang_validate_all`, `xml_yang_validate_add`, `xml_yang_validate_rpc`, `api_path2xml`, `api_path2xpath` * Added new xml functions for specific types: `xml_child_nr_notype`, `xml_child_nr_notype`, `xml_child_i_type`, `xml_find_type`.

+++ test/test_compat.sh
@@ -1,10 +1,8 @@
 #!/bin/bash
-# Startup test: Start clicon daemon in the (four) different startup modes
-# and the dbs and files are setup as follows:
-# - The example reset_state callback adds "lo" interface
-# - An extra xml configuration file starts with an "extra" interface
-# - running db starts with a "run" interface
-# - startup db starts with a "start" interface
+# Test of backward compatibility
+# 1) Load <3.9 startup/running/extra files without namespaces - ensure it returns namespaces
+#
+
 APPNAME=example
 # include err() and new() functions and creates $dir
 . ./lib.sh
@@ -34,6 +32,7 @@ cat <<EOF > $cfg
 
 EOF
 
+
 run(){
     mode=$1
     expect=$2
@@ -41,7 +40,7 @@ run(){
     dbdir=$dir/db
     cat <<EOF > $dbdir
 <config>
-   <interfaces xmlns="urn:ietf:params:xml:ns:yang:ietf-interfaces">
+   <interfaces>
       <interface>
          <name>run</name>
          <type>ex:eth</type>
@@ -53,7 +52,7 @@ EOF
 
     cat <<EOF > $dbdir
 <config>
-   <interfaces xmlns="urn:ietf:params:xml:ns:yang:ietf-interfaces">
+   <interfaces>
       <interface>
          <name>startup</name>
          <type>ex:eth</type>
@@ -65,7 +64,7 @@ EOF
 
     cat <<EOF > $dir/config
 <config>
-   <interfaces xmlns="urn:ietf:params:xml:ns:yang:ietf-interfaces">
+   <interfaces>
       <interface>
          <name>extra</name>
          <type>ex:eth</type>
@@ -74,21 +73,28 @@ EOF
 </config>
 EOF
 
-    # kill old backend (if any)
-    new "kill old backend"
-    sudo clixon_backend -zf $cfg
-    if [ $? -ne 0 ]; then
-	err
-    fi
+    new "test params: -f $cfg -s $mode -c $dir/config"
 
-    new "start backend  -f $cfg -s $mode -c $dir/config"
-    sudo $clixon_backend -f $cfg -s $mode -c $dir/config
-    if [ $? -ne 0 ]; then
-	err
+    if [ $BE -ne 0 ]; then
+	new "kill old backend"
+	sudo clixon_backend -zf $cfg
+	if [ $? -ne 0 ]; then
+ 	    err
+	fi
+    
+	new "start backend  -f $cfg -s $mode -c $dir/config"
+	sudo $clixon_backend -f $cfg -s $mode -c $dir/config
+	if [ $? -ne 0 ]; then
+	    err
+	fi
     fi
 
     new "Check $mode"
     expecteof "$clixon_netconf -qf $cfg" 0 '<rpc><get-config><source><running/></source></get-config></rpc>]]>]]>' "^<rpc-reply>$expect</rpc-reply>]]>]]>$"
+
+    if [ $BE -eq 0 ]; then
+	exit # BE
+    fi
 
     new "Kill backend"
     # Check if premature kill
@@ -101,10 +107,9 @@ EOF
     if [ $? -ne 0 ]; then
 	err "kill backend"
     fi
-}
 
-run init    '<data/>'
-run none    '<data><interfaces xmlns="urn:ietf:params:xml:ns:yang:ietf-interfaces"><interface><name>run</name><type>ex:eth</type><enabled>true</enabled></interface></interfaces></data>'
+} # run 
+
 run running '<data><interfaces xmlns="urn:ietf:params:xml:ns:yang:ietf-interfaces"><interface><name>extra</name><type>ex:eth</type><enabled>true</enabled></interface><interface><name>lo</name><type>ex:loopback</type><enabled>true</enabled></interface><interface><name>run</name><type>ex:eth</type><enabled>true</enabled></interface></interfaces></data>'
 run startup '<data><interfaces xmlns="urn:ietf:params:xml:ns:yang:ietf-interfaces"><interface><name>extra</name><type>ex:eth</type><enabled>true</enabled></interface><interface><name>lo</name><type>ex:loopback</type><enabled>true</enabled></interface><interface><name>startup</name><type>ex:eth</type><enabled>true</enabled></interface></interfaces></data>'
